+++ Common/src/main/java/fuzs/puzzleslib/api/client/core/v1/context/RenderTypesContext.java
@@ -1,8 +1,11 @@
 package fuzs.puzzleslib.api.client.core.v1.context;
 
+import fuzs.puzzleslib.api.client.renderer.v1.chunk.ChunkSectionLayer;
 import net.minecraft.client.renderer.RenderType;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.material.Fluid;
+
+import java.util.Objects;
 
 /**
  * Register custom {@link RenderType}s for blocks and fluids.
@@ -12,13 +15,31 @@ import net.minecraft.world.level.material.Fluid;
 public interface RenderTypesContext<T> {
 
     /**
-     * Register a <code>renderType</code> for an <code>object</code>
+     * Register a {@link RenderType}.
      *
-     * @param renderType the {@link RenderType} for <code>object</code>
-     * @param objects    object types supporting render type, either {@link Block} or {@link Fluid}
+     * @param object     the object
+     * @param renderType the render type
      */
+    void registerRenderType(T object, RenderType renderType);
+
+    /**
+     * Register a {@link ChunkSectionLayer}.
+     *
+     * @param object            the object
+     * @param chunkSectionLayer the chunk section layer
+     */
+    default void registerChunkRenderType(T object, ChunkSectionLayer chunkSectionLayer) {
+        this.registerRenderType(object, chunkSectionLayer.renderType);
+    }
+
+    @Deprecated
     @SuppressWarnings("unchecked")
-    void registerRenderType(RenderType renderType, T... objects);
+    default void registerRenderType(RenderType renderType, T... objects) {
+        Objects.requireNonNull(objects, "objects is null");
+        for (T object : objects) {
+            this.registerRenderType(object, renderType);
+        }
+    }
 
     /**
      * Allows for retrieving the {@link RenderType} that has been registered for an object type.