+++ Common/src/main/java/fuzs/puzzleslib/impl/client/core/context/BlockRenderTypesContextImpl.java
@@ -1,8 +1,8 @@
 package fuzs.puzzleslib.impl.client.core.context;
 
-import com.google.common.base.Preconditions;
 import fuzs.puzzleslib.api.client.core.v1.ClientAbstractions;
 import fuzs.puzzleslib.api.client.core.v1.context.RenderTypesContext;
+import fuzs.puzzleslib.api.client.renderer.v1.RenderTypeHelper;
 import net.minecraft.client.renderer.RenderType;
 import net.minecraft.world.level.block.Block;
 
@@ -11,14 +11,10 @@ import java.util.Objects;
 public final class BlockRenderTypesContextImpl implements RenderTypesContext<Block> {
 
     @Override
-    public void registerRenderType(RenderType renderType, Block... blocks) {
+    public void registerRenderType(Block block, RenderType renderType) {
+        Objects.requireNonNull(block, "block is null");
         Objects.requireNonNull(renderType, "render type is null");
-        Objects.requireNonNull(blocks, "blocks is null");
-        Preconditions.checkState(blocks.length > 0, "blocks is empty");
-        for (Block block : blocks) {
-            Objects.requireNonNull(block, "block is null");
-            ClientAbstractions.INSTANCE.registerRenderType(block, renderType);
-        }
+        RenderTypeHelper.registerRenderType(block, renderType);
     }
 
     @Override