+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v21.1.48-1.21.1] - 2025-05-27
+
+### Added
+
+- Backport `ChunkSectionLayer` from Minecraft 1.21.11
+- Backport missing `BlockSetFamily` related methods in `AbstractRecipeProvider` & `AbstractTagProvider`
+
 ## [v21.1.47-1.21.1] - 2025-05-26
 
 ### Added